Improved and refactored calls to data services Removed remote settings view
git-svn-id: https://plugins.svn.wordpress.org/hotspots/trunk@838308 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ hotspots.php
@@ -3,7 +3,7 @@
 Plugin Name: Hotspots Analytics
 Plugin URI: http://wordpress.org/extend/plugins/hotspots/
 Description: The most advanced analytics plugin for WordPress websites including heatmaps, user activity and custom event tracking.
-Version: 4.0.2
+Version: 4.0.3
 Author: Daniel Powney
 Auhtor URI: danielpowney.com
 License: GPL2
@@ -14,27 +14,29 @@ require_once dirname( __FILE__ ) . DIRECTORY_SEPARATOR . 'php' . DIRECTORY_SEPAR
 
 $ha_admin_controller = null;
 $ha_frontend_controller = null;
-if (is_admin()) {
+
+if (is_admin() && class_exists('HA_Admin_Controller')) {
 	$ha_admin_controller = new HA_Admin_Controller();
-} else {
+} else if (class_exists('HA_Frontend_Controller')) {
 	$ha_frontend_controller = new HA_Frontend_Controller();
 }
 
 // Activation and deactivation
-register_activation_hook( __FILE__, 'ha_activate_plugin');
-register_uninstall_hook( __FILE__, 'ha_uninstall_plugin' );
-register_deactivation_hook( __FILE__, 'ha_uninstall_plugin' );
 function ha_activate_plugin() {
-	if (is_admin()) {
+	if (is_admin() && class_exists('HA_Admin_Controller')) {
 		HA_Admin_Controller::activate_plugin();
 	}
 
 }
 function ha_uninstall_plugin() {
-	if (is_admin()) {
+	if (is_admin() && class_exists('HA_Admin_Controller')) {
 		HA_Admin_Controller::uninstall_plugin();
 	}
 }
+
+register_activation_hook( __FILE__, 'ha_activate_plugin');
+register_uninstall_hook( __FILE__, 'ha_uninstall_plugin' );
+//register_deactivation_hook( __FILE__, 'ha_uninstall_plugin' );
 
 // ensure a session has been started
 session_start();
